now functional but not correct
