+++ client-side/js/Components/BusinessResource.js
@@ -1,4 +1,4 @@
-export default function BusinessResource() {
+export default function BusinessResource(businessResource) {
     return `
     <div id="filteredBusinesses">
 <h1>Location: ${resources.name}</h1>
